Further massage rubocop issues
