+++ app/controllers/admin/domains/registry_lock_controller.rb
@@ -5,11 +5,9 @@ module Admin
         set_domain
         authorize! :manage, @domain
         if @domain.remove_registry_lock
-          redirect_to edit_admin_domain_url(@domain),
-                      notice: t('admin.domains.registry_lock_delete.success')
+          redirect_to edit_admin_domain_url(@domain), notice: t('.success')
         else
-          redirect_to edit_admin_domain_url(@domain),
-                      alert: t('admin.domains.registry_lock_delete.error')
+          redirect_to edit_admin_domain_url(@domain), alert: t('.error')
         end
       end
