+++ app/models/domain.rb
@@ -6,7 +6,7 @@ class Domain < ActiveRecord::Base
   include Concerns::Domain::ForceDelete
   include Concerns::Domain::Deletable
   include Concerns::Domain::Transferable
-  include Concerns::Domain::Lockable
+  include Concerns::Domain::RegistryLockable
 
   has_paper_trail class_name: "DomainVersion", meta: { children: :children_log }
